Penalized precision matrix estimation (c++)

Penalized precision matrix estimation using the graphical lasso (glasso) algorithm

GLASSOc(S, initSigma, initOmega, lam, crit_out = "avg", crit_in = "loss",
  tol_out = 1e-04, tol_in = 1e-04, maxit_out = 10000L,
  maxit_in = 10000L)

Arguments

S
pxp sample covariance matrix (denominator n).

initSigma
initialization matrix for estimated covariance matrix Sigma

initOmega
initialization matrix for Omega used to initialize the Betas

lam
tuning parameter for lasso penalty.

crit_out
criterion for convergence in outer (blockwise) loop. Criterion avg will loop until the average absolute parameter change is less than tol_out times tolerance multiple. Criterion max will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than tol_out. Defaults to avg.

crit_in
criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion loss will loop until the relative change in the objective for each response after an iteration is less than tol_in. Criterion avg will loop until the average absolute change for each response is less than tol_in times tolerance multiple. Similary, criterion max will loop until the maximum absolute change is less than tol_in times tolerance multiple. Defaults to loss.

tol_out
convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.

tol_in
convergence tolerance for inner (lasso) loop. Defaults to 1e-4.

maxit_out
maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.

maxit_in
maximum number of iterations for inner (lasso) loop. Defaults to 1e4.

Value

returns list of returns which includes:

Iterations

number of iterations.

lam

optimal tuning parameters.

Omega

estimated penalized precision matrix.

Sigma

estimated covariance matrix.

Details

For details on the implementation of 'GLASSOO', see the vignette https://mgallow.github.io/GLASSOO/.

References

Friedman, Jerome, Trevor Hastie, and Robert Tibshirani. 'Sparse inverse covariance estimation with the graphical lasso.' Biostatistics 9.3 (2008): 432-441.

Banerjee, Onureen, Ghauoui, Laurent El, and d'Aspremont, Alexandre. 2008. 'Model Selection through Sparse Maximum Likelihood Estimation for Multivariate Gaussian or Binary Data.' Journal of Machine Learning Research 9: 485-516.

Tibshirani, Robert. 1996. 'Regression Shrinkage and Selection via the Lasso.' Journal of the Royal Statistical Society. Series B (Methodological). JSTOR: 267-288.

Meinshausen, Nicolai and Buhlmann, Peter. 2006. 'High-Dimensional Graphs and Variable Selection with the Lasso.' The Annals of Statistics. JSTOR: 1436-1462.

Witten, Daniela M, Friedman, Jerome H, and Simon, Noah. 2011. 'New Insights and Faster computations for the Graphical Lasso.' Journal of Computation and Graphical Statistics. Taylor and Francis: 892-900.

Tibshirani, Robert, Bien, Jacob, Friedman, Jerome, Hastie, Trevor, Simon, Noah, Jonathan, Taylor, and Tibshirani, Ryan J. 'Strong Rules for Discarding Predictors in Lasso-Type Problems.' Journal of the Royal Statistical Society: Series B (Statistical Methodology). Wiley Online Library 74 (2): 245-266.

Ghaoui, Laurent El, Viallon, Vivian, and Rabbani, Tarek. 2010. 'Safe Feature Elimination for the Lasso and Sparse Supervised Learning Problems.' arXiv preprint arXiv: 1009.4219.

Osborne, Michael R, Presnell, Brett, and Turlach, Berwin A. 'On the Lasso and its Dual.' Journal of Computational and Graphical Statistics. Taylor and Francis 9 (2): 319-337.

Rothman, Adam. 2017. 'STAT 8931 notes on an algorithm to compute the Lasso-penalized Gausssian likelihood precision matrix estimator.'